workers/python-worker/requirements-train.txt
torch
torchvision
ultralytics
scikit-learn
onnx
onnxruntime
matplotlib
pandas
pyyaml
